frontend/requirements.txt
streamlit
requests
pandas
matplotlib
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
